Move some docker logic to a docker.js module
