+++ lib/pip.js
@@ -1,8 +1,8 @@
 const fse = require('fs-extra');
 const path = require('path');
 const {spawnSync} = require('child_process');
-const isWsl = require('is-wsl');
 const {quote} = require('shell-quote');
+const {buildImage, getBindPath} = require('./docker');
 
 /**
  * pip install the requirements to the .serverless/requirements directory
@@ -53,60 +53,18 @@ function installRequirements() {
   if (this.options.dockerizePip) {
     cmd = 'docker';
 
-    let dockerImage = this.options.dockerImage;
+    // Build docker image if required
+    let dockerImage;
     if (this.options.dockerFile) {
       this.serverless.cli.log(`Building custom docker image from ${this.options.dockerFile}...`);
-      // Build docker image from provided Dockerfile
-      const imageName = 'sls-py-reqs-custom';
-      options = [
-        'build', '-f', this.options.dockerFile, '-t', imageName, '.'
-      ];
-      const ps = spawnSync(cmd, options, {
-        'timeout': 10000,
-        'encoding': 'utf-8'
-      });
-      if (ps.error) {
-        if (ps.error.code === 'ENOENT') {
-          throw new Error('docker not found! Please install it.');
-        }
-        throw new Error(ps.error);
-      } else if (ps.status !== 0) {
-        throw new Error(ps.stderr);
-      }
-
-      // Set dockerImage option value
-      dockerImage = imageName;
+      dockerImage = buildImage(this.options.dockerFile);
+    } else {
+      dockerImage = this.options.dockerImage;
     }
-
     this.serverless.cli.log(`Docker Image: ${dockerImage}`);
 
-    // Determine os platform of docker CLI from 'docker version'
-    options = ['version', '--format', '{{with .Client}}{{.Os}}{{end}}'];
-    const ps = spawnSync(cmd, options, {'timeout': 10000, 'encoding': 'utf-8'});
-    if (ps.error) {
-      if (ps.error.code === 'ENOENT') {
-        throw new Error('docker not found! Please install it.');
-      }
-      throw new Error(ps.error);
-    } else if (ps.status !== 0) {
-      throw new Error(ps.stderr);
-    }
-
-    let bindPath;
-    const cliPlatform = ps.stdout.trim();
-    if (process.platform === 'win32') {
-      bindPath = this.servicePath.replace(/\\([^\s])/g, '/$1');
-      if (cliPlatform === 'windows') {
-        bindPath = bindPath.replace(/^\/(\w)\//i, '$1:/');
-      }
-    } else if (isWsl) {
-      bindPath = this.servicePath.replace(/^\/mnt\//, '/');
-      if (cliPlatform === 'windows') {
-        bindPath = bindPath.replace(/^\/(\w)\//i, '$1:/');
-      }
-    } else {
-      bindPath = this.servicePath;
-    }
+    // Prepare bind path depending on os platform
+    const bindPath = getBindPath(this.servicePath);
 
     options = [
       'run', '--rm',
